+++ src/index.ts
@@ -19,6 +19,11 @@ import { registerSchemaPath } from './schema'
 import { mapErrorCode, mapErrorStatus } from './error'
 import type { Context } from './context'
 
+import { composeHandler } from './compose'
+
+import { ws } from './ws'
+import type { ElysiaWSContext, ElysiaWSOptions, WSTypedSchema } from './ws'
+
 import type {
 	Handler,
 	RegisteredHook,
@@ -52,8 +57,6 @@ import type {
 	TypedWSRouteToEden
 } from './types'
 import { type TSchema } from '@sinclair/typebox'
-import { ElysiaWSContext, ElysiaWSOptions, WSTypedSchema } from './ws'
-import { composeHandler } from './compose'
 
 // @ts-ignore
 import type { Permission } from '@elysiajs/fn'
@@ -130,7 +133,7 @@ export default class Elysia<Instance extends ElysiaInstance = ElysiaInstance> {
 			method,
 			path,
 			handler,
-			hooks: mergeHook(clone(this.event), hook as RegisteredHook)
+			hooks: mergeHook({ ...this.event }, hook as RegisteredHook)
 		})
 
 		const defs = this.meta[DEFS]
@@ -476,6 +479,8 @@ export default class Elysia<Instance extends ElysiaInstance = ElysiaInstance> {
 		const instance = new Elysia<any>()
 		instance.store = this.store
 
+		if (this.wsRouter) instance.use(ws())
+
 		const sandbox = run(instance)
 
 		if (sandbox.event.request.length)
@@ -487,27 +492,36 @@ export default class Elysia<Instance extends ElysiaInstance = ElysiaInstance> {
 		this.setModel(sandbox.meta[DEFS])
 
 		Object.values(instance.routes).forEach(
-			({ method, path, handler, hooks }) => {
-				if (path === '/')
-					this._addHandler(
-						method,
-						prefix,
-						handler,
-						mergeHook(hooks, {
-							error: sandbox.event.error
-						})
+			({ method, path: originalPath, handler, hooks }) => {
+				const path =
+					originalPath === '/' ? prefix : `${prefix}${originalPath}`
+
+				const hasWsRoute = instance.wsRouter?.match('subscribe', path)
+				if (hasWsRoute) {
+					const wsRoute = instance.wsRouter!.history.find(
+						([_, wsPath]) => originalPath === wsPath
 					)
-				else
-					this._addHandler(
-						method,
-						`${prefix}${path}`,
-						handler,
-						mergeHook(hooks, {
-							error: sandbox.event.error
-						})
-					)
+					if (!wsRoute) return
+
+					return this.ws(path as any, wsRoute[2] as any)
+				}
+
+				this._addHandler(
+					method,
+					path,
+					handler,
+					mergeHook(hooks, {
+						error: sandbox.event.error
+					})
+				)
 			}
 		)
+
+		if (instance.wsRouter && this.wsRouter)
+			instance.wsRouter.history.forEach(([method, path, handler]) => {
+				if (path === '/') this.wsRouter?.add(method, prefix, handler)
+				else this.wsRouter?.add(method, `${prefix}${path}`, handler)
+			})
 
 		return this as any
 	}
@@ -554,8 +568,8 @@ export default class Elysia<Instance extends ElysiaInstance = ElysiaInstance> {
 		? Elysia<NewInstance & Instance>
 		: this {
 		const instance = new Elysia<any>()
-
 		instance.store = this.store
+		if (this.wsRouter) instance.use(ws())
 
 		const sandbox = run(instance)
 
@@ -569,6 +583,16 @@ export default class Elysia<Instance extends ElysiaInstance = ElysiaInstance> {
 
 		Object.values(instance.routes).forEach(
 			({ method, path, handler, hooks: localHook }) => {
+				const hasWsRoute = instance.wsRouter?.match('subscribe', path)
+				if (hasWsRoute) {
+					const wsRoute = instance.wsRouter!.history.find(
+						([_, wsPath]) => path === wsPath
+					)
+					if (!wsRoute) return
+
+					return this.ws(path as any, wsRoute[2] as any)
+				}
+
 				this._addHandler(
 					method,
 					path,
@@ -577,6 +601,11 @@ export default class Elysia<Instance extends ElysiaInstance = ElysiaInstance> {
 				)
 			}
 		)
+
+		if (instance.wsRouter && this.wsRouter)
+			instance.wsRouter.history.forEach(([method, path, handler]) => {
+				this.wsRouter?.add(method, path, handler)
+			})
 
 		return this as any
 	}
@@ -1296,9 +1325,11 @@ export default class Elysia<Instance extends ElysiaInstance = ElysiaInstance> {
 			path,
 			// @ts-ignore
 			(context) => {
+				console.log('Got', context.request.url)
+
 				if (
 					// @ts-ignore
-					this.server!.upgrade(context.request, {
+					this.server?.upgrade(context.request, {
 						headers:
 							typeof options.headers === 'function'
 								? options.headers(context as any)